+++ packages/stg-devtools-evm-hardhat-v2/src/tip20/schema.ts
@@ -1,4 +1,4 @@
-import { Tip20EdgeConfigSchema, Tip20NodeConfigSchema } from '@stargatefinance/stg-devtools-v2'
+import { TIP20EdgeConfigSchema, TIP20NodeConfigSchema } from '@stargatefinance/stg-devtools-v2'
 
 import {
     createOmniEdgeHardhatSchema,
@@ -6,7 +6,7 @@ import {
     createOmniNodeHardhatSchema,
 } from '@layerzerolabs/devtools-evm-hardhat'
 
-export const Tip20OmniGraphHardhatSchema = createOmniGraphHardhatSchema(
-    createOmniNodeHardhatSchema(Tip20NodeConfigSchema),
-    createOmniEdgeHardhatSchema(Tip20EdgeConfigSchema)
+export const TIP20OmniGraphHardhatSchema = createOmniGraphHardhatSchema(
+    createOmniNodeHardhatSchema(TIP20NodeConfigSchema),
+    createOmniEdgeHardhatSchema(TIP20EdgeConfigSchema)
 )