+++ packages/stg-evm-v2/devtools/config/mainnet/01/usdc-tip20-token.config.ts
@@ -2,12 +2,12 @@ import { TokenName } from '@stargatefinance/stg-definitions-v2'
 
 import { OmniGraphHardhat } from '@layerzerolabs/devtools-evm-hardhat'
 
-import { buildTip20TokenGraphMainnet } from '../utils'
+import { buildTIP20TokenGraphMainnet } from '../utils'
 
-import type { Tip20NodeConfig } from '@stargatefinance/stg-devtools-v2'
+import type { TIP20NodeConfig } from '@stargatefinance/stg-devtools-v2'
 
 const tokenName = TokenName.USDC
 
-export default async (): Promise<OmniGraphHardhat<Tip20NodeConfig, unknown>> => {
-    return buildTip20TokenGraphMainnet(tokenName)
+export default async (): Promise<OmniGraphHardhat<TIP20NodeConfig, unknown>> => {
+    return buildTIP20TokenGraphMainnet(tokenName)
 }