+++ packages/stg-evm-v2/devtools/config/mainnet/utils.ts
@@ -10,7 +10,7 @@ import buildFeeLibV1DeploymentGraph from '../utils/feelib-v1.config.utils'
 import buildMessagingGraph from '../utils/messaging.config.utils'
 import buildRewarderGraph from '../utils/rewarder.config.utils'
 import buildStakingGraph from '../utils/staking.config.utils'
-import buildTip20TokenGraph from '../utils/tip20-token.config.utils'
+import buildTIP20TokenGraph from '../utils/tip20-token.config.utils'
 import buildTreasurerGraph from '../utils/treasurer.config.utils'
 import buildUsdtTokenGraph from '../utils/usdt-token.config.utils'
 import { setStage } from '../utils/utils.config'
@@ -59,6 +59,6 @@ export function buildCircleFiatTokenGraphMainnet(tokenName: TokenName) {
     return buildCircleFiatTokenGraph(Stage.MAINNET, tokenName)
 }
 
-export function buildTip20TokenGraphMainnet(tokenName: TokenName) {
-    return buildTip20TokenGraph(Stage.MAINNET, tokenName)
+export function buildTIP20TokenGraphMainnet(tokenName: TokenName) {
+    return buildTIP20TokenGraph(Stage.MAINNET, tokenName)
 }