+++ packages/stg-evm-v2/src/messaging/CreditMessaging.sol
@@ -14,10 +14,7 @@ contract CreditMessaging is MessagingBase, CreditMessagingOptions, ICreditMessag
 
     // ---------------------------------- Only Planner ------------------------------------------
 
-    function sendCredits(
-        uint32 _dstEid,
-        TargetCreditBatch[] calldata _creditBatches
-    ) external payable virtual onlyPlanner {
+    function sendCredits(uint32 _dstEid, TargetCreditBatch[] calldata _creditBatches) external payable onlyPlanner {
         CreditBatch[] memory batches = new CreditBatch[](_creditBatches.length);
         uint256 index = 0;
         uint128 totalCreditNum = 0; // total number of credits in all batches