+++ packages/stg-evm-v2/src/tip20/StargateOFTTIP20.sol
@@ -1,13 +1,16 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity ^0.8.22;
 
-import { ITIP20Minter } from "../interfaces/ITIP20Minter.sol";
-import { StargateOFTAlt } from "../StargateOFTAlt.sol";
 import { Transfer } from "../libs/Transfer.sol";
+import { StargateOFTAlt } from "../StargateOFTAlt.sol";
+import { ITIP20 } from "../interfaces/tip20/ITIP20.sol";
+import { ITIP20RolesAuth } from "../interfaces/tip20/ITIP20RolesAuth.sol";
 
-/// @dev OFT variant for bridged stablecoin with TIP-20 and ALT endpoints.
-/// - Transfers + Burns stablecoin on inflow and mints on outflow
+/// @notice StargateOFT variant for bridged stablecoin with TIP-20 and EndpointV2Alt.
+/// @dev Messages in EndpointV2Alt chains cannot be delivered in bus mode, however it is still possible to receive them.
 contract StargateOFTTIP20 is StargateOFTAlt {
+    bytes32 internal constant DEFAULT_ADMIN_ROLE = 0;
+
     constructor(
         address _token,
         uint8 _sharedDecimals,
@@ -20,14 +23,24 @@ contract StargateOFTTIP20 is StargateOFTAlt {
         amountSD = _ld2sd(_amountLD);
         _amountLD = _sd2ld(amountSD); // remove dust
         Transfer.safeTransferTokenFrom(token, _from, address(this), _amountLD);
-        ITIP20Minter(token).burn(_amountLD);
+        ITIP20(token).burn(_amountLD);
     }
 
     /// @dev Mint stablecoin to the receiver on outflow
     /// @dev TIP-20 mint function implementations do not return a boolean.
     function _outflow(address _to, uint256 _amountLD) internal virtual override returns (bool success) {
-        try ITIP20Minter(token).mint(_to, _amountLD) {
+        try ITIP20(token).mint(_to, _amountLD) {
             success = true;
         } catch {} // solhint-disable-line no-empty-blocks
+    }
+
+    /// @notice Transfer the TIP-20 admin role to a new owner.
+    /// @dev Grants DEFAULT_ADMIN_ROLE to `_newOwner` and renounces it for the current admin.
+    /// @dev It mimics the transfer ownership functionality for the ERC20 tokens with roles.
+    /// @param _newOwner The account to receive the admin role.
+    function transferTokenOwnership(address _newOwner) external virtual override onlyOwner {
+        // grant the role to the new owner and renounce it to remove it from current
+        ITIP20RolesAuth(token).grantRole(DEFAULT_ADMIN_ROLE, _newOwner);
+        ITIP20RolesAuth(token).renounceRole(DEFAULT_ADMIN_ROLE);
     }
 }