+++ packages/stg-evm-v2/test/unitest/stargate/StargateOftTIP20.t.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.0;
 
 import { StargateOftTest, IMockStargate } from "./StargateOft.t.sol";
-import { TIP20Token } from "../../../src/mocks/TIP20Token.sol";
+import { TIP20 } from "../../../src/mocks/tip20/TIP20.sol";
 import { TokenMessagingAlt } from "../../../src/messaging/TokenMessagingAlt.sol";
 import { StargateOFTTIP20 } from "../../../src/tip20/StargateOFTTIP20.sol";
 import { LzUtil } from "../../layerzero/LzUtil.sol";
@@ -13,7 +13,9 @@ import { OFTLimit, OFTFeeDetail, OFTReceipt, SendParam, MessagingReceipt, Messag
 import { AddressCast } from "@layerzerolabs/lz-evm-protocol-v2/contracts/libs/AddressCast.sol";
 
 contract StargateOftTIP20Test is StargateOftTest {
-    TIP20Token public tip20Token;
+    bytes32 internal constant DEFAULT_ADMIN_ROLE = bytes32(0);
+    bytes32 internal constant ISSUER_ROLE = keccak256("ISSUER_ROLE");
+    TIP20 public tip20Token;
     AltFeeTokenMock public feeToken;
     bytes32 internal constant MOCK_GUID = bytes32(uint(1));
 
@@ -28,9 +30,6 @@ contract StargateOftTIP20Test is StargateOftTest {
     }
 
     function test_ReceiveTokenTaxi(uint32 _amountInSD, uint64 _nonce, bytes memory _composeMsg) public {
-        // Allow OFT to mint on receive
-        tip20Token.transferOwnership(address(stargate));
-
         // 1. Set up a _composeMsg test case.
         uint256 amountInLD = stargate.sdToLd(_amountInSD);
         Origin memory origin = Origin({ srcEid: 2, sender: AddressCast.toBytes32(ALICE), nonce: _nonce });
@@ -43,8 +42,20 @@ contract StargateOftTIP20Test is StargateOftTest {
         assertEq(afterBalance, before + amountInLD);
     }
 
+    function test_TransferTokenOwnership() public {
+        tip20Token.grantRole(DEFAULT_ADMIN_ROLE, address(stargate));
+
+        assertTrue(tip20Token.hasRole(address(stargate), DEFAULT_ADMIN_ROLE));
+
+        address newOwner = address(this);
+        MockStargateOFTTIP20(address(stargate)).transferTokenOwnership(newOwner);
+
+        assertTrue(tip20Token.hasRole(newOwner, DEFAULT_ADMIN_ROLE));
+        assertFalse(tip20Token.hasRole(address(stargate), DEFAULT_ADMIN_ROLE));
+    }
+
     function _setUpStargate() internal override {
-        tip20Token = new TIP20Token("TIP20Token", "T2T");
+        tip20Token = new TIP20("TIP20Token", "T2T", "USD", TIP20(address(0)), address(this), address(this));
         feeToken = new AltFeeTokenMock();
 
         stargate = new MockStargateOFTTIP20(
@@ -53,9 +64,11 @@ contract StargateOftTIP20Test is StargateOftTest {
             LzUtil.deployEndpointV2Alt(LOCAL_EID, address(this), address(feeToken)),
             address(this)
         );
+        tip20Token.grantRole(ISSUER_ROLE, address(stargate));
     }
 
     function _deal(address _to, uint256 _amount, uint256 _fee) internal override {
+        vm.prank(address(stargate));
         tip20Token.mint(_to, _amount);
         // ALT endpoints use an ERC20 fee token instead of ETH; fund and approve it.
         if (_fee > 0) {