+++ packages/stg-evm-sdk-v2/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stargatefinance/stg-evm-sdk-v2
 
+## 1.2.7
+
+### Patch Changes
+
+- 39b37ae: Configured and deployed wETH on Redbelly
+
+## 1.2.6
+
+### Patch Changes
+
+- 927c8ea: Adding redbelly chain
+- eb73e8a: Adding Redbelly deployment
+
 ## 1.2.5
 
 ### Patch Changes